# src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"LobbyResponse": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"players": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"maxPlayers": {
						"type": "number",
						"format": "double"
					},
					"isActive": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"_id",
					"name",
					"code",
					"players",
					"maxPlayers",
					"isActive",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateLobbyRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"maxPlayers": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"JoinLobbyRequest": {
				"properties": {
					"playerName": {
						"type": "string"
					}
				},
				"required": [
					"playerName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LeaveLobbyRequest": {
				"properties": {
					"playerName": {
						"type": "string"
					}
				},
				"required": [
					"playerName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessMsg": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChallengeDifficulty": {
				"type": "string",
				"enum": [
					"Muito fácil",
					"Fácil",
					"Médio",
					"Avançado",
					"Difícil"
				]
			},
			"TestCaseInput": {
				"properties": {
					"input": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "number",
								"format": "double"
							},
							{
								"additionalProperties": false,
								"type": "object"
							},
							{
								"items": {
									"anyOf": [
										{
											"type": "string"
										},
										{
											"type": "number",
											"format": "double"
										}
									]
								},
								"type": "array"
							}
						]
					},
					"expectedOutput": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "number",
								"format": "double"
							},
							{
								"additionalProperties": false,
								"type": "object"
							},
							{
								"items": {
									"anyOf": [
										{
											"type": "string"
										},
										{
											"type": "number",
											"format": "double"
										}
									]
								},
								"type": "array"
							}
						]
					}
				},
				"required": [
					"input",
					"expectedOutput"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ChallengeResponse": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"creator": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"difficulty": {
						"$ref": "#/components/schemas/ChallengeDifficulty"
					},
					"categories": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"description": {
						"type": "string"
					},
					"testCases": {
						"items": {
							"$ref": "#/components/schemas/TestCaseInput"
						},
						"type": "array"
					}
				},
				"required": [
					"_id",
					"title",
					"creator",
					"createdAt",
					"difficulty",
					"categories",
					"description",
					"testCases"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorResponse": {
				"properties": {
					"error": {
						"type": "string"
					}
				},
				"required": [
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateChallengeRequest": {
				"properties": {
					"title": {
						"type": "string"
					},
					"creator": {
						"type": "string"
					},
					"difficulty": {
						"$ref": "#/components/schemas/ChallengeDifficulty"
					},
					"categories": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"description": {
						"type": "string"
					},
					"testCases": {
						"items": {
							"$ref": "#/components/schemas/TestCaseInput"
						},
						"type": "array"
					}
				},
				"required": [
					"title",
					"creator",
					"difficulty",
					"categories",
					"description",
					"testCases"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateChallengeRequest": {
				"properties": {
					"title": {
						"type": "string"
					},
					"difficulty": {
						"$ref": "#/components/schemas/ChallengeDifficulty"
					},
					"categories": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"description": {
						"type": "string"
					},
					"testCases": {
						"items": {
							"$ref": "#/components/schemas/TestCaseInput"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "Codex API",
		"version": "1.0.0",
		"description": "API documentation for Codex - Backend lobby and match logic",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "vitorluuiz"
		}
	},
	"paths": {
		"/api/lobby/create": {
			"post": {
				"operationId": "CreateLobby",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LobbyResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Lobby"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateLobbyRequest"
							}
						}
					}
				}
			}
		},
		"/api/lobby": {
			"get": {
				"operationId": "GetAllLobbies",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/LobbyResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Lobby"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/lobby/{code}": {
			"get": {
				"operationId": "GetLobbyByCode",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LobbyResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Lobby"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "code",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteLobby",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessMsg"
								}
							}
						}
					}
				},
				"tags": [
					"Lobby"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "code",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/lobby/{code}/join": {
			"post": {
				"operationId": "JoinLobby",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LobbyResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Lobby"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "code",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/JoinLobbyRequest"
							}
						}
					}
				}
			}
		},
		"/api/lobby/{code}/leave": {
			"post": {
				"operationId": "LeaveLobby",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LobbyResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Lobby"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "code",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LeaveLobbyRequest"
							}
						}
					}
				}
			}
		},
		"/api/challenges": {
			"post": {
				"operationId": "CreateChallenge",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ChallengeResponse"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Challenges"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateChallengeRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllChallenges",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ChallengeResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Challenges"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/challenges/difficulties": {
			"get": {
				"operationId": "GetAvailableDifficulties",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"type": "string"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Challenges"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/challenges/{id}": {
			"get": {
				"operationId": "GetChallengeById",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ChallengeResponse"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Challenges"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateChallenge",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ChallengeResponse"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Challenges"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateChallengeRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteChallenge",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/SuccessMsg"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Challenges"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/challenges/creator/{creator}": {
			"get": {
				"operationId": "GetChallengesByCreator",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ChallengeResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Challenges"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "creator",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/challenges/difficulty/{difficulty}": {
			"get": {
				"operationId": "GetChallengesByDifficulty",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ChallengeResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Challenges"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "difficulty",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/challenges/category/{category}": {
			"get": {
				"operationId": "GetChallengesByCategory",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ChallengeResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Challenges"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "category",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}